+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/CarController.h
@@ -24,7 +24,7 @@ public:
 
     float GetClutchCmd() const;
 
-    float GetLightCmd() const;
+    int GetLightCmd() const;
 
     void ShowIntervention(InterventionAction p_element) const;
